# src/main/resources/api.yml
---
swagger: '2.0'
info:
  version: 1.0.0
  title: Wallet API
securityDefinitions:
  basicAuth:
    type: basic
paths:
  /wallets:
    post:
      summary: Create a new wallet
      operationId: createWallet
      security:
       - basicAuth: []
      produces:
      - application/json
      responses:
        201:
          description: Created
          headers:
            Location:
              type: string
              description: Wallet location
          schema:
            $ref: "#/definitions/wallet"
  /wallets/{id}:
    parameters:
    - name: id
      in: path
      type: integer
      format: int64
      required: true
    get:
      summary: Get a wallet
      operationId: getWallet
      security:
       - basicAuth: []
      produces:
      - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/wallet"
  /wallets/{id}/transactions:
    parameters:
    - name: id
      in: path
      type: integer
      format: int64
      required: true
    post:
      summary: Create transaction
      operationId: createTransaction
      security:
       - basicAuth: []
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: transaction
        in: body
        schema:
          $ref: "#/definitions/transaction"
        required: true
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/wallet"
        403:
          description: Not enough funds
          schema:
            $ref: "#/definitions/error"
          examples:
            application/json:
              message: Not enough funds
definitions:
  error:
    properties:
      message:
        type: string
    required:
    - message
  wallet:
    properties:
      balance:
        type: number
        multipleOf: 0.01
    required:
    - balance
  transaction:
    properties:
      amount:
        type: number
        multipleOf: 0.01
      category:
        type: string
        enum:
        - WAGER
        - PAYOUT
    required:
    - amount